+++ app/src/main/java/com/pict/pbl/medswift/MainActivity.kt
@@ -13,8 +13,10 @@ import androidx.compose.foundation.text.KeyboardOptions
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.filled.ArrowForward
 import androidx.compose.material.icons.filled.Email
+import androidx.compose.material.icons.filled.Lock
 import androidx.compose.material3.*
 import androidx.compose.runtime.*
+import androidx.compose.runtime.livedata.observeAsState
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.draw.drawBehind
@@ -24,12 +26,15 @@ import androidx.compose.ui.text.input.KeyboardType
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.constraintlayout.compose.ConstraintLayout
+import androidx.lifecycle.MutableLiveData
+import com.pict.pbl.medswift.login.LoginManager
 import com.pict.pbl.medswift.ui.theme.MedSwiftTheme
 
 class MainActivity : ComponentActivity() {
 
     val blue1 = Color( 0xFF2055f5 )
     val blue2 = Color( 0xFF6185f2 )
+    private val isLoginButtonEnabled = MutableLiveData( false )
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
@@ -56,24 +61,15 @@ class MainActivity : ComponentActivity() {
                 drawCircle( color = blue1 , radius = 512.dp.toPx() )
                 drawCircle( color = blue2 , radius = 320.dp.toPx() )
                 drawCircle( color = Color.White , radius = 240.dp.toPx() )
-            }
-                ) {
+            }) {
+            // TODO: Add app icon here, with Image( ... )
             EmailId( modifier = Modifier
                 .fillMaxWidth()
                 .padding(start = 32.dp, end = 32.dp) )
             Password( modifier = Modifier
                 .fillMaxWidth()
                 .padding(start = 32.dp, end = 32.dp) )
-            Button(
-                onClick = {  } ,
-                modifier = Modifier
-                    .padding( 16.dp ) ,
-                colors = ButtonDefaults.buttonColors(
-                    containerColor = blue1
-                )) {
-                Icon( imageVector = Icons.Default.ArrowForward , contentDescription = "Login" )
-                Text(text = "Login")
-            }
+            LoginButton(modifier = Modifier.padding(16.dp))
         }
     }
 
@@ -81,17 +77,22 @@ class MainActivity : ComponentActivity() {
     @Composable
     private fun EmailId( modifier: Modifier ){
         var email by remember{ mutableStateOf("") }
+        // TODO: Add singleLine = true attribute here
         OutlinedTextField(
             modifier = modifier ,
             value = email,
-            onValueChange = { it -> email = it } ,
+            onValueChange = { it ->
+                email = it
+                isLoginButtonEnabled.value = LoginManager.checkEmailAddress( it )
+            } ,
             placeholder = { Text("Enter your EmailID" , color = Color.Black) } ,
             leadingIcon = { Icon( imageVector = Icons.Default.Email , contentDescription = "Email Address" ) }  ,
             colors = TextFieldDefaults.outlinedTextFieldColors(
                 focusedBorderColor = Color.Blue,
                 focusedLeadingIconColor = Color.Blue,
                 containerColor = Color.White
-            )
+            ) ,
+            keyboardOptions = KeyboardOptions( keyboardType = KeyboardType.Email )
         )
     }
 
@@ -99,16 +100,36 @@ class MainActivity : ComponentActivity() {
     @Composable
     private fun Password( modifier : Modifier ){
         var password by remember{ mutableStateOf("") }
+        // TODO: Add singleLine = true attribute here
+        // TODO: Implement 'Tap to show password' button here, refer
+        //       https://stackoverflow.com/a/66998457/13546426
         OutlinedTextField(
             modifier = modifier,
             value = password,
             onValueChange = { it -> password = it} ,
+            placeholder = { Text("Password" , color = Color.Black) } ,
+            leadingIcon = { Icon( imageVector = Icons.Default.Lock , contentDescription = "Email Address" ) } ,
             colors = TextFieldDefaults.outlinedTextFieldColors(
                 focusedBorderColor = Color.Blue,
                 containerColor = Color.White
             ),
             keyboardOptions = KeyboardOptions( keyboardType = KeyboardType.Password )
         )
+    }
+
+    @Composable
+    private fun LoginButton( modifier: Modifier ) {
+        val isEnabled by isLoginButtonEnabled.observeAsState()
+        Button(
+            onClick = {  } ,
+            modifier = modifier,
+            enabled = isEnabled ?: false ,
+            colors = ButtonDefaults.buttonColors(
+                containerColor = blue1
+            )) {
+            Icon( imageVector = Icons.Default.ArrowForward , contentDescription = "Login" )
+            Text(text = "Login")
+        }
     }
 
 }