+++ app/src/main/java/com/pict/pbl/medswift/login/LoginManager.kt
@@ -1,0 +1,16 @@
+package com.pict.pbl.medswift.login
+
+class LoginManager {
+
+    companion object {
+        fun checkEmailAddress( emailAddress: String ) : Boolean {
+            // TODO: Check if email address is well-formed, use RegEx
+            return true
+        }
+    }
+
+    fun validateUser( emailAddress : String , password : String ) {
+        // TODO: Add user validation here
+    }
+
+}