+++ CHANGELOG.md
@@ -1,1 +1,4 @@
-- Started project
+- Created separate `Composable` for `LoginButton` in MainActivity.kt
+- Added icon to `Password` composable
+- Added `keyboardType = KeyboardType.Email` to `EmailId` composable
+- Added check for valid email address in `LoginManager.kt`